refactor(app): change List<unsigned long long> to List<size_t> for better type consistency

+++ shaykhraziev.bulat/S1/app.cpp
@@ -11,11 +11,11 @@ shaykhraziev::List< shaykhraziev::NamedSeq > shaykhraziev::readInput(std::istrea
   {
     NamedSeq seq;
     seq.name = name;
-    List< unsigned long long >::iterator numTail = seq.nums.before_begin();
+    List< size_t >::iterator numTail = seq.nums.before_begin();
     skipSpaces(in);
     while (!isLineEnd(in))
     {
-      unsigned long long x = 0;
+      size_t x = 0;
       in >> x;
       numTail = seq.nums.insert_after(numTail, x);
       skipSpaces(in);
@@ -26,9 +26,9 @@ shaykhraziev::List< shaykhraziev::NamedSeq > shaykhraziev::readInput(std::istrea
   return seqs;
 }
 
-bool shaykhraziev::addChecked(unsigned long long a, unsigned long long b, unsigned long long& result)
+bool shaykhraziev::addChecked(size_t a, size_t b, size_t& result)
 {
-  if (b > std::numeric_limits< unsigned long long >::max() - a)
+  if (b > std::numeric_limits< size_t >::max() - a)
   {
     return false;
   }
@@ -53,26 +53,26 @@ void shaykhraziev::printNames(std::ostream& out, const List< NamedSeq >& seqs)
 
 int shaykhraziev::printSequences(std::ostream& out, std::ostream& err, const List< NamedSeq >& seqs)
 {
-  List< List< unsigned long long >::const_iterator > iters;
+  List< List< size_t >::const_iterator > iters;
   {
-    List< List< unsigned long long >::const_iterator >::iterator itTail = iters.before_begin();
+    List< List< size_t >::const_iterator >::iterator itTail = iters.before_begin();
     for (LCIter< NamedSeq > s = seqs.cbegin(); s != seqs.cend(); ++s)
     {
       itTail = iters.insert_after(itTail, s->nums.cbegin());
     }
   }
 
-  List< unsigned long long > sums;
-  List< unsigned long long >::iterator sumTail = sums.before_begin();
+  List< size_t > sums;
+  List< size_t >::iterator sumTail = sums.before_begin();
 
   bool anyRow = true;
   while (anyRow)
   {
     anyRow = false;
     bool firstVal = true;
-    unsigned long long rowSum = 0;
+    size_t rowSum = 0;
 
-    LIter< List< unsigned long long >::const_iterator > it = iters.begin();
+    LIter< List< size_t >::const_iterator > it = iters.begin();
     LCIter< NamedSeq > s = seqs.cbegin();
 
     while (it != iters.end())
@@ -84,10 +84,10 @@ int shaykhraziev::printSequences(std::ostream& out, std::ostream& err, const Lis
         {
           out << ' ';
         }
-        unsigned long long val = **it;
+        size_t val = **it;
         out << val;
         firstVal = false;
-        unsigned long long newSum = 0;
+        size_t newSum = 0;
         if (!addChecked(rowSum, val, newSum))
         {
           err << "overflow\n";
@@ -110,7 +110,7 @@ int shaykhraziev::printSequences(std::ostream& out, std::ostream& err, const Lis
   if (!sums.empty())
   {
     bool firstSum = true;
-    for (LCIter< unsigned long long > si = sums.cbegin(); si != sums.cend(); ++si)
+    for (LCIter< size_t > si = sums.cbegin(); si != sums.cend(); ++si)
     {
       if (!firstSum)
       {
