+++ shaykhraziev.bulat/S1/app.hpp
@@ -4,17 +4,18 @@
 #include "list.hpp"
 #include <string>
 #include <iostream>
+#include <cstddef>
 
 namespace shaykhraziev
 {
   struct NamedSeq
   {
     std::string name;
-    List< unsigned long long > nums;
+    List< size_t > nums;
   };
 
   List< NamedSeq > readInput(std::istream& in);
-  bool addChecked(unsigned long long a, unsigned long long b, unsigned long long& result);
+  bool addChecked(size_t a, size_t b, size_t& result);
   void printNames(std::ostream& out, const List< NamedSeq >& seqs);
   int printSequences(std::ostream& out, std::ostream& err, const List< NamedSeq >& seqs);
 }
